+++ packages/zp-plugins/src/index.ts
@@ -1,6 +1,8 @@
 export * from './config'
 export * from './express'
+export * from './mongodb'
 
 export { default as hooks } from './hooks'
 export { default as config } from './config'
 export { default as express } from './express'
+export { default as mongodb } from './mongodb'