+++ packages/zp-plugins/CHANGELOG.md
@@ -1,9 +1,13 @@
 # zapnode-plugins
 
+## 0.0.10
+
+### Patch Changes
+
+- add mongodb plugin
+
 ## 0.0.9
 
 ### Patch Changes
 
 - add mongodb service to zapnode-data
-- Updated dependencies
-  - zapnode@0.0.9