+++ playground/hello/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hello
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zapnode-plugins@0.0.10
+
 ## 1.0.1
 
 ### Patch Changes